+++ CHANGELOG.md
@@ -11,21 +11,31 @@ rm src/utils/validation.js
 rm src/utils/__tests__/schema.test.js
 rm src/utils/__tests__/validation.test.js
 rm -r src/utils/__fixtures__
+yarn add @bedrockio/yada
+yarn add @bedrockio/logger
 yarn add @bedrockio/model
 ```
+
+Replace all references to `src/utils/schema.js` and `src/utils/validation.js` to `@bedrockio/model`.
 
 In the model definitions:
 
 - Move `search` fields into a nested `fields`.
 - Rename `readScopes` to `readAccess`.
 - Rename `writeScopes` to `writeAccess`.
+- Remove `skipValidation`.
+- Change `Mixed` type to `Object`.
 
-```json
+```js
 {
   "profile" {
     "type": "String",
     "readScopes": "none",
     "writeScopes": "none",
+    "skipValidation": true,
+  },
+  "object": {
+    "type": "Mixed"
   },
   "search": ["firstName", "lastName"]
 }
@@ -33,12 +43,15 @@ In the model definitions:
 
 should be migrated to:
 
-```json
+```js
 {
   "profile" {
     "type": "String",
     "readAccess": "none",
     "writeAccess": "none",
+  },
+  "object": {
+    "type": "Object"
   },
   "search": {
     "fields": ["firstName", "lastName"]